Change the tests to a much simpler 2d+skill system.
I was unhappy with the probabilistic properties of the old skill
system. Primarily by the fact that the difference between unskilled
and highly skilled students was so huge: more than three-fold. Further,
the system required a lot of rolling and rerolling in play. The new
system is an attempt to shift most of the complication to advancement,
and make the actual rolling much simpler.